Code Style and Lint
The source adopted the LLVM code style, because we have a lot of tools
to help this standard.

Use clang-format, when installed, to check the modified C++ files before
commit. And use the shellcheck to verify the bash modified files.

The default build.cmake script already link the git pre-commit hook.